apps/api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
SQLAlchemy==2.0.35
pydantic==2.12.5
email-validator==2.2.0
PyJWT==2.9.0
httpx==0.27.2
pytest==8.3.3
